+++ MDPapp/mobile/src/main/java/tudelft/mdp/locationTracker/NetworkScanService.java
@@ -98,7 +98,7 @@ public class NetworkScanService extends Service {
                 NetworkInfoObject networkInfo = new NetworkInfoObject(
                         result.SSID,
                         result.BSSID,
-                        result.level);
+                        (double) result.level);
                 scanResult.add(networkInfo);
             }
             getNewScanResults();
